Extract the share logic to base class
